refactor: extract struct ClientId
