+++ src/udp/connection/mod.rs
@@ -1,4 +1,5 @@
 
 pub mod byte_array_32;
+pub mod client_id;
 pub mod connection_id;
 pub mod timestamp;
